Weird loop is happening

+++ src/main/scala/evaluator/objects/IntegerObject.scala
@@ -2,7 +2,7 @@ package evaluator.objects
 
 import evaluator.{Anything, ObjectType}
 
-case class IntegerObject(value: Int) extends Anything {
+case class IntegerObject(value: BigInt) extends Anything {
   override def objectType: ObjectType = ObjectType.Integer
   override def inspect: String = value.toString
 }
